+++ utoipa/src/openapi.rs
@@ -163,17 +163,10 @@ impl OpenApi {
         serde_json::to_string_pretty(self)
     }
 
-    /// Converts this [`OpenApi`] to YAML String. This method essentially calls [`serde_norway::to_string`] method.
+    /// Converts this [`OpenApi`] to YAML String. This method essentially calls [`yaml_serde::to_string`].
     #[cfg(feature = "yaml")]
     #[cfg_attr(doc_cfg, doc(cfg(feature = "yaml")))]
-    pub fn to_yaml(&self) -> Result<String, serde_norway::Error> {
-        serde_norway::to_string(self)
-    }
-
-    /// Converts this [`OpenApi`] to YAML String. This method essentially calls [`yaml_serde::to_string`].
-    #[cfg(feature = "yaml_serde")]
-    #[cfg_attr(doc_cfg, doc(cfg(feature = "yaml")))]
-    pub fn to_yaml_serde(&self) -> Result<String, yaml_serde::Error> {
+    pub fn to_yaml(&self) -> Result<String, yaml_serde::Error> {
         yaml_serde::to_string(self)
     }